Implement countBombsNearbyAllCells function
Implemented the 'countBombsNearbyAllCells' function in the Board class. This function iterates over all cells on the board and counts the number of bombs adjacent to each cell. The resulting count is then stored in the corresponding cell on the board.

Note: This commit does not yet include the logic for eliminating adjacent cells with a count of '0'. That will be implemented in a subsequent commit.